refactor: create runner/cleanup package
