+++ pkg/runner/cleanup/skip.go
@@ -1,6 +1,6 @@
-package runner
+package cleanup
 
-func skipDelete(config bool, test *bool, step *bool) bool {
+func Skip(config bool, test *bool, step *bool) bool {
 	if step != nil {
 		return *step
 	}
